agent/pyproject.toml
[project]
name = "gtm-quest-agent"
version = "0.1.0"
description = "GTM.quest AI Agent - Pydantic AI with CopilotKit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "fastapi",
    "starlette",
    "pydantic-ai-slim[ag-ui]",
    "pydantic-ai-slim[google]",
    "python-dotenv",
    "psycopg2-binary",
    "httpx",
]
